Graph2, ConstantDB and SimpleMappedGraph
Graph2 is a new base interface for a graph. It lets the neighbors be
fetched over a network, by wrapping them in a future. In addition, the
underlying node ids are now longs.

ConstantDB is a port of D. J. Bernstein's excellent cdb package
(http://cr.yp.to/cdb.html), with a few changes. Unlike cdb, there is
no limit on the total database size.  A single entry must be smaller
than 2GB. Additionally, when creating the database the input keys must
be provided in sorted order.

SimpleMappedGraph is a graph implementation that sits on top of
ConstantDB and implements Graph2.

WIP

Initial version of graph2

WIP

WIP

WIP

More docs